included onnx export module
i full-on stole the export module from a feature branch in brevitas, about 100 commits behind
this means that that version does not include conv1d, so i'm using "globally" installed brevitas
+ locally shoved module for onnx export. now have to pick a specific size, pad the whole
dataset, pass the size to the finn export and it should be working.
hopefully the Conv1D won't be SO different (unlikely) that the export func does not know how
to deal with it
not sure how the padding thing is done. i guess the BucketIterator already knows how to do it,
but setting a really high batch size seems very stupid. also, now i get why the example network
in brevitas was fixed-size :/