restricted all sequences to length 3000, small change to export
i knew padding and non-variable sequences were already implemented,
and i found it in torchtext.Field. then the export generated a useless
input to trace the model, so i switched that out for the first example
in the dataset. present situation is: `RuntimeError: Tried to trace
<__torch__.brevitas.core.quant.RescalingIntQuant object at 0x55b84e4f5130>
but it is not part of the active trace. Modules that are called during
a trace must be registered as submodules of the thing being traced.`
there seems to be some support online:
- is it possible the input is supposed to be quantized?
- should try pytorch-nightly
- why are we using brevitas when real full-fledged libs exist (caffe
seems to be one of them)
- i don't know how quantization really works, or what's really going
wrong, or if brevitas would be able to trace modules that were only
included in later versions. so, not sure where to go from here